Image generation moved to new User_Avatar_Handler
